[*] moved PomoSettingsModal to components/modals
